requirements.txt
black==21.11b1
click==8.0.3
Flask==2.0.2
Flask-Login==0.5.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
regex==2021.11.10
tomli==1.2.2
typing_extensions==4.0.0
Werkzeug==2.0.2
python-dotenv
requests
authlib
six